+++ test/runtests.jl
@@ -16,6 +16,7 @@ isok(r) = r.status == 200
         "chunking.jl",
         "utils.jl",
         "client.jl",
+        "sse.jl",
         # "download.jl",
         "multipart.jl",
         "parsemultipart.jl",